+++ visualiser/generators/okitGenerator.py
@@ -33,6 +33,8 @@ class OCIGenerator(object):
         self.create_sequence = []
         self.run_variables = {}
         self.data_output = []
+        self.connection_provider = []
+        self.metadata = []
         # Assign passed values to local
         self.template_dir = template_dir
         self.output_dir = output_dir
@@ -61,6 +63,24 @@ class OCIGenerator(object):
         self.jinja2_variables["author"] = __author__
         self.jinja2_variables["copyright"] = __copyright__
         self.jinja2_variables["okit_version"] = self.OKIT_VERSION
+        self.jinja2_variables["deployment_platform"] = self.visualiser_json.get('metadata', {}).get('platform', 'oci')
+    
+    def addStandardResourceVariables(self, resource={}):
+        standardisedName = self.standardiseResourceName(resource.get('resource_name', resource['display_name']))
+        self.jinja2_variables['resource_name'] = standardisedName
+        self.jinja2_variables['output_name'] = self.standardiseOutputName(resource['display_name'])
+        # ---- Read Only
+        self.jinja2_variables['read_only'] = resource.get('read_only', False)
+        # ---- Id
+        self.jinja2_variables["ocid"] = self.formatJinja2Value(resource['id'])
+        # --- Required
+        # ---- Compartment Id
+        self.jinja2_variables["compartment_id"] = self.formatJinja2IdReference(self.standardiseResourceName(self.id_name_map[resource['compartment_id']]))
+        # ---- Display Name
+        self.addJinja2Variable("display_name", resource["display_name"], standardisedName)
+        # ---- Tags
+        self.renderTags(resource)
+
 
     def get(self, artifact_type, id):
         artifact = {};
@@ -76,6 +96,12 @@ class OCIGenerator(object):
         elif not os.path.isdir(self.output_dir):
             logger.error('Output directory {0:s} is not a directory'.format(self.output_dir))
 
+    def getProvider(self):
+        return self.connection_provider
+
+    def getMetadata(self):
+        return self.metadata
+
     def getRenderedMain(self):
         return self.create_sequence
 
@@ -86,6 +112,9 @@ class OCIGenerator(object):
         return self.create_sequence
 
     def writeFiles(self):
+        pass
+
+    def toJson(self):
         pass
 
     def formatJinja2Variable(self, variable_name):