+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 ## Version 0.31.0
 **Release Date**: 19th January 2022
 ### Features
+1. Add Terraform view which will display the generated Terraform (First Release)
 ### Bug Fixes
+1. Fix Standardised resource name issue when display_name contain non-alphanumeric characters.
 
 
 ## Version 0.30.0